requirements.txt
# ============================================================================
# Single authoritative dependency specification for MolBridge
#
# This file consolidates what was previously split across requirements.txt and
# requirements-prod.txt. Optional / heavier integrations are marked clearly; you
# can comment them out if not needed (e.g. minimal deployment).
# To regenerate a locked snapshot for reproducible builds:
#   pip install -r requirements.txt && pip freeze --exclude-editable > requirements.lock
# ============================================================================

# Core Streamlit and web framework
streamlit>=1.32.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Biological data processing
biopython>=1.83
biopandas>=0.4.0
biotite>=0.37.0

# Scientific computing
numpy>=1.24.0
scipy>=1.12.0
pandas>=2.2.0
scikit-learn>=1.4.0

# 3D structure analysis
py3Dmol>=2.0.4
MDAnalysis>=2.7.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
streamlit-lottie>=0.0.5
networkx>=3.2.0  # For force-directed interaction network layout
rapidfuzz>=3.6.1   # For fuzzy search in command palette

# Report generation
reportlab>=4.0.0
openpyxl>=3.1.0
python-pptx>=0.6.23
Pillow>=10.2.0
fpdf2>=2.7.0

# File handling and utilities
requests>=2.31.0
httpx>=0.27.0
aiofiles>=23.2.0
python-multipart>=0.0.9

# Progress tracking
tqdm>=4.66.0
stqdm>=0.0.5

# Caching and performance
diskcache>=5.6.0
joblib>=1.3.0

# Configuration and logging
pydantic>=2.6.0
pydantic-settings>=2.0.0  # Needed for BaseSettings in Pydantic v2+
python-dotenv>=1.0.0
loguru>=0.7.0
PyYAML>=6.0.0  # For scenario profile YAML loading

# System / performance monitoring (optional)
psutil>=5.9.0

# Testing & benchmarking (optional / dev)
pytest>=8.0.0          # OPTIONAL: test framework
pytest-benchmark>=4.0.0  # OPTIONAL: performance benchmarks

# Optional acceleration
numba>=0.59.0          # OPTIONAL: JIT acceleration (enable with MOLBRIDGE_USE_NUMBA=1)

# Optional ML/AI integrations (comment out if unused)
openai>=1.12.0      # OPTIONAL: AI-assisted summarization
anthropic>=0.18.0   # OPTIONAL: Alternative LLM backend

# LaTeX export (optional; requires system pandoc if full PDF pipeline used)
pypandoc>=1.12      # OPTIONAL: Rich LaTeX / docx conversions
